switch from mostly-lexicographical permutation enumeration to a more efficient transposition-based one.
Essentially, we decompose each permutation into a list of transpositions of increasing index: e.g. `P = (4 a4)(3 a3)(2 a2)(1 a1)` for S5, where `ai >= i` for all `i`.  This is easy to compute in both directions: the index-to-permutation case is trivial since it's just composing the transpositions, and the permutation-to-index case can be done incrementally by locking in each element in place, starting from the left.  To ensure index parity corresponds to permutation parity, we count the number of non-zero elements before the final term (which is always a 0 or 1).  If that number is odd, then we invert it before adding, making it effectively the full parity rather than an ordinary term.